# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Fast Note Sync Service HTTP API.",
        "title": "Fast Note Sync Service HTTP API",
        "contact": {
            "name": "Haierkeys",
            "url": "https://github.com/haierkeys",
            "email": "haierkeys@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/api/file": {
            "get": {
                "security": [
                    {
                        "UserAuthToken": []
                    }
                ],
                "description": "Get raw binary data of an attachment by path, supports strong cache control",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Get attachment content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "isRecycle",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pathHash",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "vault",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UserAuthToken": []
                    }
                ],
                "description": "Permanently delete a specific attachment record and its physical file",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Delete attachment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pathHash",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "vault",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.FileDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/file/info": {
            "get": {
                "security": [
                    {
                        "UserAuthToken": []
                    }
                ],
                "description": "Get attachment metadata (FileDTO) by path",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Get attachment info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "isRecycle",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pathHash",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "vault",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.FileDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/files": {
            "get": {
                "security": [
                    {
                        "UserAuthToken": []
                    }
                ],
                "description": "Get attachment list for current user with pagination, search, filter, and sort support",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Get file list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "isRecycle",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting field: mtime(default), ctime, path",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting order: desc(default), asc",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "vault",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.FileDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/share/file": {
            "get": {
                "security": [
                    {
                        "ShareAuthToken": []
                    }
                ],
                "description": "Get raw binary data of a specific attachment via share token",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Share"
                ],
                "summary": "Get shared attachment content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth Token",
                        "name": "Share-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.Res": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {},
                "data": {},
                "details": {},
                "message": {},
                "status": {
                    "type": "boolean"
                },
                "vault": {}
            }
        },
...
        "dto.FileDTO": {
            "type": "object",
            "properties": {
                "contentHash": {
                    "type": "string"
                },
                "ctime": {
                    "type": "integer"
                },
                "lastTime": {
                    "type": "integer"
                },
                "mtime": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "pathHash": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ShareAuthToken": {
            "type": "apiKey",
            "name": "Share-Token",
            "in": "header"
        },
        "UserAuthToken": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}